separate package for dom
in case we plan to publish it later on
