+++ dom/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "dom",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Native dom implementation layer over react-native",
+  "license": "MIT",
+  "exports": {
+    ".": {
+      "browser": "./dist/dom.module.js",
+      "umd": "./dist/dom.umd.js",
+      "import": "./dist/dom.mjs",
+      "require": "./dist/dom.js"
+    }
+  },
+  "main": "dist/dom.js",
+  "module": "dist/dom.module.js",
+  "source": "src/index.js",
+  "peerDependencies": {
+    "react": ">=17",
+    "react-native": ">=0.65"
+  }
+}
