--- src/dom/index.js
@@ -1,6 +1,0 @@
-export * from "./bridge.js"
-export * from "./element.js"
-export * from "./shared.js"
-export * from "./node.js"
-export * from "./renderer.js"
-export * from "./utils.js"
