--- src/index.js
@@ -1,2 +1,0 @@
-export * from "./dom"
-export * from "./preact"
